Minimized changes to trunk at revision [4168] (this branch initially contained modifications from ARM not yet incorporated on trunk. It now contains only the modifications to trunk@[4168] relevant to the ADC)
git-svn-id: https://epos.lisha.ufsc.br/svn/epos2@4191 86035c5f-d46c-4136-a2d5-fe63f1d4a782

+++ branches/merge/ticket_29_emote3_adc/include/utility/malloc.h
@@ -62,7 +62,7 @@ private:
 
 public:
     static void * alloc(unsigned int bytes, const EPOS::Color & allocator) {
-        assert((unsigned int)allocator <= COLORS);
+        assert(allocator <= COLORS);
         return _heap[allocator]->alloc(bytes);
     }
